Tweaking of the slowing down of gesture scrolling; removed debug output.
In the prev. version, diagonal scrolling would become horizontal or
vertical after some time; this is fixed now.

(To see the issue, choose a high magnification so that one can scroll
far in any direction; then do a gesture to scroll in a direction e.g.
somewhere between diagonal and vertical. After a short time, it scrolls
completely vertically.)